requirements.txt
streamlit==1.28.1
spacy==3.7.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.24.3
pdfplumber==0.10.3
plotly==5.17.0
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0